+++ CHANGELOG.md
@@ -13,5 +13,6 @@
   * Fixed pipeline path is none error ([#16](https://github.com/KIT-IAI/pywatts-pipeline/issues/16))
   * Fixed summary step crashing ([#25](https://github.com/KIT-IAI/pywatts-pipeline/issues/25))
   * Pipeline result and step buffer are reset after training ([#35](https://github.com/KIT-IAI/pywatts-pipeline/issues/35))
-  * Fix the start value in the online execution ([#37](https://github.com/KIT-IAI/pywatts-pipeline/issues/33))
+  * Fix the start value in the online execution ([#37](https://github.com/KIT-IAI/pywatts-pipeline/issues/37))
+  * Train estimator if computation mode is refit and estimator is not fitted yet ([#33](https://github.com/KIT-IAI/pywatts-pipeline/issues/33))